[bundleInit.sh] Feature: Now check if the feeders and data exist

+++ gatling-liferay-portlet/src/main/webapp/WEB-INF/classes/gatling/bundleInit.sh
@@ -9,12 +9,18 @@ dataDir=user-files/bodies/liferay
 # Copies all the scala files
 mkdir -p $GATLING_HOME/$simulationsDir
 mkdir -p $GATLING_HOME/$processesDir
-mkdir -p $GATLING_HOME/$dataDir
 cp $simulationsDir/*.scala $GATLING_HOME/$simulationsDir
 cp $processesDir/* $GATLING_HOME/$processesDir
-cp $dataDir/* $GATLING_HOME/$dataDir
+
+# Copies all the http bodies
+if [ -e $dataDir ]; then
+  mkdir -p $GATLING_HOME/$dataDir
+  cp $dataDir/* $GATLING_HOME/$dataDir
+fi
 
 # Copies all the feeders
-mkdir -p $GATLING_HOME/$feedersDir
-cp $feedersDir/* $GATLING_HOME/$feedersDir
+if [ -e $feedersDir ]; then
+  mkdir -p $GATLING_HOME/$feedersDir
+  cp $feedersDir/* $GATLING_HOME/$feedersDir
+fi
